DataFrame cleanups
