+++ dask/dataframe/__init__.py
@@ -1,9 +1,9 @@
-from .core import (DataFrame, Series, Index, _Frame, concat, map_partitions,
+from .core import (DataFrame, Series, Index, _Frame, map_partitions,
         repartition)
 from .io import (read_csv, from_array, from_bcolz, from_array, from_bcolz,
                  from_pandas, from_dask_array, from_castra, read_hdf)
 from .optimize import optimize
-from .multi import merge
+from .multi import merge, concat
 from .rolling import (rolling_count, rolling_sum, rolling_mean, rolling_median,
                       rolling_min, rolling_max, rolling_std, rolling_var,
                       rolling_skew, rolling_kurt, rolling_quantile, rolling_apply,
