+++ dask/dataframe/tests/test_dataframe.py
@@ -12,7 +12,8 @@ import dask
 from dask.async import get_sync
 from dask.utils import raises, ignoring
 import dask.dataframe as dd
-from dask.dataframe.core import (concat, repartition_divisions, _loc,
+
+from dask.dataframe.core import (repartition_divisions, _loc,
         _coerce_loc_index, aca, reduction, _concat, _Frame)
